+++ src/main.c
@@ -4,13 +4,13 @@
 #include "response.h"
 
 int main() {
-    int solution = 0;
+    unsigned long solution = 0;
     const AocResponse response = problem10_solve(&solution);
     if (!aoc_is_success(&response)) {
         aoc_display_response(&response);
         return 1;
     }
 
-    printf("The solution to the problem is: %d\n", solution);
+    printf("The solution to the problem is: %zu\n", solution);
     return 0;
 }
